test(torghut): split DSPy runtime hash coverage
